requirements.txt
asgiref==3.7.2
Django==4.2.1
pip==23.1.2
setuptools==67.8.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.40.0
djangorestframework-jwt
djangorestframework
Pillow
